webkit/data/layout_test_results/v8/LayoutTests/fast/parser/script-tag-with-trailing-slash-expected.txt

CONSOLE MESSAGE: line 2: Uncaught SyntaxError: Unexpected token <
You should not see the text "inside" the following script tag: This is the text inside the script tag.
